requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
beautifulsoup4==4.12.3
google-generativeai==0.4.0
python-dotenv==1.0.0
pillow==10.2.0
pydantic==2.5.3
